Lots of changes!
Too long since I last pushed - did some work on Ahoy, but most of the
changes are in obfuscate/parse, obfuscate looks like it won't work out
(got stuck with left-side-recurrence) but it was a learning experience
that has enabled me to work on parse (which is problem-free, and looks
like it should be extensible enough to handle any js.) The match project
is also in here, it has some cool potential but I think it needs a
rewrite. Aighty let's push this shiz.