BracketedValue and BinaryOp complete!
Could use some abstraction in order to minimize code. E.g. BinaryOp
looks a lot like Indirecting (both consume the head and recurse on the
tail). StatementBlock (coming soon) will probably look similar as well.
Will probably be important for the "parse" method to return unconsumed
input instead of just an index indicating the number of items consumed.
This will be because "parse" may want to modify certain tokens, for
instance if it is only able to consume a portion of the last token, and
wants to indicate that some of that token still remains unconsumed.