Possibly fixed a big stupid bug
Turns out that Session was only gaining a reference to the app the first
time it was loaded. Heroku probably queries the app from some random IP,
causing an anonymous heroku session to initialize. After that the
Session never needs to initialize the app, but it was mistakenly also
not pointing its "app" property to the app.